+++ app/domain/output_document.rb
@@ -74,32 +74,8 @@ class OutputDocument
     appointment_summary.guider_name
   end
 
-  def pages_to_render
-    case variant
-    when 'tailored'
-      circumstances = %i(continue_working unsure leave_inheritance
-                         wants_flexibility wants_security wants_lump_sum
-                         poor_health).select do |c|
-        appointment_summary.public_send("#{c}?".to_sym)
-      end
-
-      [:cover_letter, :introduction, circumstances, :other_information].flatten
-    when 'generic'
-      %w(cover_letter introduction generic_guidance other_information)
-    when 'other'
-      %w(ineligible)
-    end
-  end
-
-  def stylesheet(filename)
-    css = Sass.compile(ERB.new(
-      File.read(Rails.root.join('app', 'assets', 'stylesheets', filename))
-    ).result(binding))
-    "<style>\n#{css}\n</style>".html_safe
-  end
-
   def html
-    ERB.new(template).result(binding)
+    HTMLRenderer.new(self).render
   end
 
   def csv
@@ -107,17 +83,6 @@ class OutputDocument
   end
 
   private
-
-  def template
-    [:header, pages_to_render, :footer].flatten.reduce('') do |result, section|
-      result << template_for(section) << "\n\n"
-    end
-  end
-
-  def template_for(section)
-    File.read(
-      Rails.root.join('app', 'templates', "output_document_#{section}.html.erb"))
-  end
 
   def to_currency(number)
     return '' if number.blank?
